refactor: split component out
