+++ src/components/ResultWindow.tsx
@@ -1,0 +1,7 @@
+type ResultWindowProps = {
+  wpm: number;
+};
+
+export const ResultWindow = ({ wpm }: ResultWindowProps) => {
+  return <div>WPM: {wpm}</div>;
+};
